pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["imnotdev25 <85677268+imnotdev25@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
motor = "^3.5.1"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.2"
pillow = "^10.4.0"
aiofiles = "^24.1.0"
google-generativeai = "^0.7.2"
pyrogram = "^2.0.106"
python-dotenv = "^1.0.1"
uvloop = "^0.19.0"
openai = "^1.35.14"
tgcrypto = "^1.2.5"
speedtest-cli = "^2.1.3"
telegraph = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
